refactor(auth): SecurityConfig 를 presentation 계층으로 이동해 의존 방향 정리
- SecurityConfig 가 presentation 의 JwtAuthenticationFilter/핸들러를 import 해 infrastructure→presentation 역방향 의존이 있었다.
- SecurityConfig 를 presentation/common/security 로 옮겨 presentation→domain 한 방향 의존만 유지한다.
- JwtProperties 스캔은 infrastructure 신규 JwtConfig 에서 담당하도록 분리한다.
