+++ src/main/java/com/readum/infrastructure/auth/config/JwtConfig.java
@@ -1,0 +1,9 @@
+package com.readum.infrastructure.auth.config;
+
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
+import org.springframework.context.annotation.Configuration;
+
+@Configuration
+@EnableConfigurationProperties(JwtProperties.class)
+public class JwtConfig {
+}
